Store device credential values exactly as sent, and fold partial updates onto their columns in core
A device credential's value was trimmed when it was saved, while the
password a device presented was compared untrimmed, so an MQTT password
with leading or trailing whitespace could never authenticate. Credential
values are now stored exactly as sent; only an empty value stores no
secret.

Partial updates on nullable columns now go through folds in core that take
and return the column type: ApplyToNullString, ApplyToNullSecret,
ApplyToNullInt64 and ApplyToIntPtr. A field the caller did not name comes
back untouched. Before, the shared spelling re-trimmed stored text, and
notification policies narrowed stored integers through int32, on exactly
that branch. The value rules live once, in a small leaf package (sqlnull)
that both the create paths and the folds use, so core's graphql package
does not pull the database drivers into every module that serves a
schema.

Reading a stored integer wider than a GraphQL Int is now refused with a
resolver error instead of silently wrapped, for notification policy
intervals and for the tenant governance values user-management serves.

An identity's first and last name and an AI provider's endpoint are
nullable in the model, as their columns already were, and appended
migrations convert the empty strings earlier releases stored to NULL. A
tenant tier's "no color" becomes NULL through an appended migration that
drops the column's NOT NULL and default. Reads render a stored empty
string as null too, so a row written by a previous-release pod during a
rolling upgrade reads the same. With that, the private fold that wrote
empty strings in place of NULL is deleted, and so are the per-service
copies of the integer helpers.

The partial-update test harness gains OptionalBoolField and
RequiredStringListField. It now drives the [] state on required lists and
requires it to be refused, and it fails any three-state list field on a
request type whose Field cannot send [], where before it skipped one.

+++ backend/services/user-management/graphql/admin_tenant_settings_test.go
@@ -72,9 +72,9 @@ func TestEffectiveSettingsReportsTierAndDelta(t *testing.T) {
 	assert.Equal(t, float64(5000), *ingest.Rate().Override())
 
 	assert.Equal(t, string(iam.SourceOverride), ingest.Burst().Source())
-	assert.EqualValues(t, 9000, *ingest.Burst().Value())
-	assert.EqualValues(t, 4000, *ingest.Burst().Tier())
-	assert.EqualValues(t, 9000, *ingest.Burst().Override())
+	assert.EqualValues(t, 9000, *i32(ingest.Burst().Value()))
+	assert.EqualValues(t, 4000, *i32(ingest.Burst().Tier()))
+	assert.EqualValues(t, 9000, *i32(ingest.Burst().Override()))
 
 	// An un-overridden dimension reports the tier as the winner and carries no
 	// delta — nil, not zero: "no exception" and "an exception of 0" are opposite
@@ -84,7 +84,7 @@ func TestEffectiveSettingsReportsTierAndDelta(t *testing.T) {
 	assert.Equal(t, float64(200), *outbound.Rate().Value())
 	assert.Equal(t, float64(200), *outbound.Rate().Tier())
 	assert.Nil(t, outbound.Rate().Override())
-	assert.Nil(t, outbound.Burst().Override())
+	assert.Nil(t, i32(outbound.Burst().Override()))
 }
 
 // TestEffectiveSettingsNamesThePlatformDefaultWithoutInventingIt pins the honest
@@ -110,7 +110,7 @@ func TestEffectiveSettingsNamesThePlatformDefaultWithoutInventingIt(t *testing.T
 		assert.Nil(t, row.Rate().Override(), name)
 
 		assert.Equal(t, string(iam.SourcePlatformDefault), row.Burst().Source(), name)
-		assert.Nil(t, row.Burst().Value(), name)
+		assert.Nil(t, i32(row.Burst().Value()), name)
 	}
 }
